requirements.txt
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
google-api-core==2.12.0
google-auth==2.23.2
google-cloud-texttospeech==2.14.2
googleapis-common-protos==1.60.0
grpcio==1.59.0
grpcio-status==1.59.0
idna==3.4
proto-plus==1.22.3
protobuf==4.24.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
requests==2.31.0
rsa==4.9
urllib3==2.0.6
